refactor RandAugment and AutoAugment
